+++ flow-typed/modules/youboralib.js
@@ -1,0 +1,4 @@
+// @flow
+declare module 'youboralib' {
+  declare module.exports: any;
+}