+++ karma.conf.js
@@ -10,43 +10,30 @@ const isMacOS = /^darwin/.test(process.platform);
 const customLaunchers = {
   Chrome_travis_ci: {
     base: 'Chrome',
-    flags: ['--no-sandbox']
+    flags: ['--no-sandbox', '--autoplay-policy=no-user-gesture-required']
   }
 };
 
-module.exports = function (config) {
+module.exports = function(config) {
   let karmaConf = {
     logLevel: config.LOG_INFO,
-    browsers: [
-      'Chrome',
-      'Firefox'
-    ],
+    browsers: ['Chrome', 'Firefox'],
     concurrency: 1,
     singleRun: true,
     colors: true,
-    frameworks: [
-      'mocha'
-    ],
+    frameworks: ['mocha'],
     files: [
-      'test/setup/karma.js', {
+      'test/setup/karma.js',
+      {
         pattern: 'src/assets/audios.mp4',
         included: false
       }
     ],
     preprocessors: {
-      'src/**/*.js': [
-        'webpack',
-        'sourcemap'
-      ],
-      'test/setup/karma.js': [
-        'webpack',
-        'sourcemap'
-      ]
+      'src/**/*.js': ['webpack', 'sourcemap'],
+      'test/setup/karma.js': ['webpack', 'sourcemap']
     },
-    reporters: [
-      'progress',
-      'coverage'
-    ],
+    reporters: ['mocha', 'coverage'],
     webpack: webpackConfig,
     webpackServer: {
       noInfo: true
@@ -61,9 +48,7 @@ module.exports = function (config) {
 
   if (process.env.TRAVIS) {
     karmaConf.customLaunchers = customLaunchers;
-    karmaConf.browsers = [
-      'Chrome_travis_ci'
-    ];
+    karmaConf.browsers = ['Chrome_travis_ci'];
   } else {
     if (isWindows) {
       karmaConf.browsers.push('IE');