+++ src/index.js
@@ -1,5 +1,5 @@
 // @flow
-import {registerPlugin} from 'playkit-js'
+import {registerPlugin} from '@playkit-js/playkit-js';
 import {Plugin} from './youbora'
 
 declare var __VERSION__: string;
@@ -8,17 +8,15 @@ declare var __NAME__: string;
 const VERSION = __VERSION__;
 const NAME = __NAME__;
 
-export {VERSION, NAME};
-export {Plugin};
+export {Plugin, VERSION, NAME};
 
 /**
  * The plugin name.
  * @type {string}
  * @const
  */
-const pluginName = "youbora";
+const pluginName = 'youbora';
 /**
  * Register the plugin in player's registry.
  */
 registerPlugin(pluginName, Plugin);
-